Move some blacklist handling code to a much better location
Handle weird non-standard cases where chips are repeated, like:
  chip 1 1 core 1
  chip 1 1 core 2
